[Refactor] Next.js 14 버전으로 마이그레이션 (#15)
* Chore: next14 패키지 설치

* Chore: eslint 설정 Next 프로젝트에 맞게 변경, ignore와 기존 파일 삭제

* Chore: tsconfig 수정

* Chore: eslint설정 수정하면서 package.json 정리

- eslint-config-next를 사용하면서 필요없어진 eslint-plugin-react, hooks 패키지를 제거하고 vite 관련 패키지도 제거

* Chore: gitignore 파일 수정

* Chore: eslintrc.json 수정

* Chore: 마이그레이션 과정에서 vite 관련 파일 삭제

* Chore: 마이그레이션 과정에서 파일구조 변경에 따른 파일이동 및 삭제

* Chore: prettier설정 파일의 형식 변경

* Chore: tsconfig에 incremental 옵션 추가

* Refactor: vite에서 next로 변경함에 따른 코드 변경

- RootLayout내용을 next에 맞게 변경
- page.tsx에서 Main컴포넌트를 클라이언트 컴포넌트로 가져와서 사용하도록 구현

* Chore: vanilla-extract의 nextjs 패키지 설치

* Style: lint 설정에 따른 코드 스타일 일괄변경

* Chore: next.config.mjs에 vanilla-extract 설정 추가

* Chore: @types/gsap 설치

* Fix: 마이그레이션 후 새로고침 시 Main컴포넌트 런타임 오류 발생 문제 수정

- 처음 로드 후 새로고침을 할 때 Main컴포넌트의 ScrollTrigger config설정 부분에서 런타임 오류(참조오류)가 발생했습니다
- CSR에서 Next방식으로 바뀌면서 발생한 문제로, 설정부분을 useLayoutEffect로 마운트 후 실행되도록 옮겨서 해결했습니다

* Refactor: Introduction 컴포넌트 img를 next/image의 Image로 변경

* Chore: @gsap/react 패키지 업그레이드, sharp 패키지 설치

* Chore: 이미지 경로에 대한 옵션을 next.config.mjs에 추가

* Refactor: img태그를 next/image의 Image 태그로 변경
