+++ .eslintignore
@@ -1,0 +1,6 @@
+next.config.js
+/next.config.js
+node_modules
+postcss.config.js
+tailwind.config.js
+prettier.config.js
