+++ package.json
@@ -4,36 +4,36 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "vite preview",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "eslint ./src/**/*.{js,jsx,ts,tsx} --cache",
+    "lint:fix": "eslint ./src/**/*.{js,jsx,ts,tsx} --cache --fix",
     "format": "prettier --write ./src"
   },
   "dependencies": {
-    "@gsap/react": "^2.0.2",
+    "@gsap/react": "^2.1.0",
     "@vanilla-extract/css": "^1.14.0",
     "gsap": "^3.12.5",
+    "next": "^14.1.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "sharp": "^0.33.2"
   },
   "devDependencies": {
+    "@types/gsap": "^3.0.0",
     "@types/node": "^20.10.5",
     "@types/react": "^18.2.43",
     "@types/react-dom": "^18.2.17",
     "@typescript-eslint/eslint-plugin": "^6.15.0",
     "@typescript-eslint/parser": "^6.15.0",
-    "@vanilla-extract/vite-plugin": "^3.9.3",
-    "@vitejs/plugin-react": "^4.2.1",
+    "@vanilla-extract/next-plugin": "^2.3.6",
     "eslint": "^8.56.0",
+    "eslint-config-next": "14.0.4",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-jsx-a11y": "^6.8.0",
-    "eslint-plugin-prettier": "^5.1.0",
-    "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.5",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-prettier": "^5.1.3",
     "prettier": "3.1.1",
-    "typescript": "^5.3.3",
-    "vite": "^5.0.8"
+    "typescript": "^5.3.3"
   }
 }
