+++ src/app/MainPage.css.ts
@@ -1,8 +1,8 @@
-import { style } from "@vanilla-extract/css";
+import { style } from '@vanilla-extract/css';
 
 export const mainPageHeading = style({
-  color: "red",
-  fontSize: "50px",
-  fontWeight: "bold",
-  margin: "0",
+  color: 'red',
+  fontSize: '50px',
+  fontWeight: 'bold',
+  margin: '0',
 });
