+++ src/app/RootLayout.css.ts
@@ -1,9 +1,9 @@
-import { style } from "@vanilla-extract/css";
+import { style } from '@vanilla-extract/css';
 
 export const RootLayout = style({
-  overflow: "hidden",
+  overflow: 'hidden',
 });
 
 export const MainContainer = style({
-  padding: "0 170px",
+  padding: '0 170px',
 });
