+++ src/app/_component/checkout/Checkout.css.ts
@@ -1,0 +1,11 @@
+import { style } from '@vanilla-extract/css';
+
+export const CheckoutContainer = style({
+  position: 'relative',
+  width: '100%',
+  height: '100vh',
+  display: 'flex',
+  flexDirection: 'column',
+  justifyContent: 'space-between',
+  alignItems: 'center',
+});
