+++ src/app/layout.tsx
@@ -1,15 +1,28 @@
-import * as styles from "./RootLayout.css";
-import Footer from "./component/common/footer/Footer";
-// import HeaderNav from "./component/common/nav/HeaderNav";
+import type { Metadata } from 'next';
 
-const RootLayout: React.FC<{ children: React.ReactNode }> = ({ children }) => {
-  return (
-    <main className={styles.RootLayout}>
-      {/* <HeaderNav /> */}
-      <section className={styles.MainContainer}>{children}</section>
-      <Footer />
-    </main>
-  );
+import { Inter } from 'next/font/google';
+import '@/app/RootLayout.css';
+
+const inter = Inter({
+  subsets: ['latin'],
+  display: 'swap',
+});
+
+export const metadata: Metadata = {
+  title: '이원형 멤버쉽 사이트',
+  description: '이원형에 대한 개인적인 후원을 통해 멤버쉽 혜택을 받아보세요.',
 };
 
-export default RootLayout;
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="ko">
+      <body className={`${inter.className} antialiased`}>
+        <main>{children}</main>
+      </body>
+    </html>
+  );
+}
