+++ src/assets/MembershipBenefits.ts
@@ -1,32 +1,32 @@
 export const membershipBenefitsData = [
   {
-    title: "라이트 혜택",
-    description: "저렴한 가격으로 맛보기 기능을 제공합니다",
+    title: '라이트 혜택',
+    description: '저렴한 가격으로 맛보기 기능을 제공합니다',
     benefits: [
-      { id: "1", benefit: "기상, 취침 시간 알림권" },
-      { id: "2", benefit: "일 1회 게임 동원권" },
-      { id: "3", benefit: "일 1회 롤 서렌 거부권" },
-      { id: "4", benefit: "3끼 사진 제공" },
+      { id: '1', benefit: '기상, 취침 시간 알림권' },
+      { id: '2', benefit: '일 1회 게임 동원권' },
+      { id: '3', benefit: '일 1회 롤 서렌 거부권' },
+      { id: '4', benefit: '3끼 사진 제공' },
     ],
   },
   {
-    title: "일반 혜택",
-    description: "기본 가격으로 표준혜택을 제공합니다",
+    title: '일반 혜택',
+    description: '기본 가격으로 표준혜택을 제공합니다',
     benefits: [
-      { id: "1", benefit: "일반 혜택 1" },
-      { id: "2", benefit: "일반 혜택 2" },
-      { id: "3", benefit: "일반 혜택 3" },
-      { id: "4", benefit: "일반 혜택 4" },
+      { id: '1', benefit: '일반 혜택 1' },
+      { id: '2', benefit: '일반 혜택 2' },
+      { id: '3', benefit: '일반 혜택 3' },
+      { id: '4', benefit: '일반 혜택 4' },
     ],
   },
   {
-    title: "프리미엄 혜택",
-    description: "일반 혜택에 더해 여러 추가 기능을 제공합니다",
+    title: '프리미엄 혜택',
+    description: '일반 혜택에 더해 여러 추가 기능을 제공합니다',
     benefits: [
-      { id: "1", benefit: "프리미엄 혜택 1" },
-      { id: "2", benefit: "프리미엄 혜택 2" },
-      { id: "3", benefit: "프리미엄 혜택 3" },
-      { id: "4", benefit: "프리미엄 혜택 4" },
+      { id: '1', benefit: '프리미엄 혜택 1' },
+      { id: '2', benefit: '프리미엄 혜택 2' },
+      { id: '3', benefit: '프리미엄 혜택 3' },
+      { id: '4', benefit: '프리미엄 혜택 4' },
     ],
   },
 ];
