+++ src/assets/PaymentData.ts
@@ -1,17 +1,17 @@
 export const paymentData = [
   {
-    title: "라이트",
-    image: "images/lite.png",
+    title: '라이트',
+    image: 'images/lite.png',
     가격: 990,
   },
   {
-    title: "일반",
-    image: "images/normal.png",
+    title: '일반',
+    image: 'images/normal.png',
     가격: 2990,
   },
   {
-    title: "프리미엄",
-    image: "images/premium.png",
+    title: '프리미엄',
+    image: 'images/premium.png',
     가격: 4990,
   },
 ];
