+++ src/assets/SponsorsReviewData.ts
@@ -1,12 +1,12 @@
 export const sponsorsReviewData = [
   {
-    name: "근박",
-    review: "이길수있어요. 힘내봐요. 서렌치지마요.",
+    name: '근박',
+    review: '이길수있어요. 힘내봐요. 서렌치지마요.',
     rating: 5,
   },
   {
-    name: "기박",
-    review: "돈이 아깝고 머리가 아파요. 0점이 없어서 1점 드릴게요.",
+    name: '기박',
+    review: '돈이 아깝고 머리가 아파요. 0점이 없어서 1점 드릴게요.',
     rating: 1,
   },
 ];
